Backend/requirements.txt
gunicorn==23.0.0
Flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.0.1
PyPDF2==3.0.1
requests==2.32.3
beautifulsoup4==4.13.3
playwright==1.54.0
faiss-cpu==1.11.0
langchain==0.3.25
langchain-core==0.3.60
langchain-community==0.3.24
langchain-text-splitters==0.3.8
langchain-openai>=0.2.0,<0.3.0
openai>=1.55.0,<2.0.0
tiktoken==0.9.0
